+++ src/tpc_monitor_lbw.cpp
@@ -40,7 +40,7 @@ std::vector<int32_t>::const_iterator LowBwTpcMonitor::deserialize(std::vector<in
     it = Serializer<LowBwTpcMonitor>::deserialize_tuple(member_tuple(), begin, end);
 
     // Ensure there's enough data for the bins
-    if (std::distance(it, end) < NUM_CHARGE_CHANNELS) {
+    if (static_cast<size_t>(std::distance(it, end)) < NUM_CHARGE_CHANNELS) {
         throw std::runtime_error("Deserialization failed: not enough data for Number charge channels.");
     }
